Refactor: move elastic APIs into a folder.
