+++ examples/pytorch/mnist/lightning_train.py
@@ -21,7 +21,7 @@ from torch.nn import functional as F
 from torch.utils.data import DataLoader
 from torchvision import transforms
 
-from dlrover.trainer.torch.elastic_sampler import ElasticDistributedSampler
+from dlrover.trainer.torch.elastic.sampler import ElasticDistributedSampler
 
 
 class UpdateDataStepCallback(pl.Callback):
